Arreglado nada

+++ src/java/org/sghweb/controllers/ActividadJpaController.java
@@ -5,19 +5,22 @@
 package org.sghweb.controllers;
 
 import java.io.Serializable;
+import javax.persistence.Query;
+import javax.persistence.EntityNotFoundException;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
+import org.sghweb.jpa.Turno;
+import java.util.ArrayList;
 import java.util.List;
 import javax.annotation.Resource;
 import javax.naming.Context;
 import javax.naming.InitialContext;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
-import javax.persistence.Query;
-import javax.persistence.EntityNotFoundException;
 import javax.persistence.Persistence;
 import javax.persistence.PersistenceUnit;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
 import javax.transaction.UserTransaction;
+import org.sghweb.controllers.exceptions.IllegalOrphanException;
 import org.sghweb.controllers.exceptions.NonexistentEntityException;
 import org.sghweb.controllers.exceptions.PreexistingEntityException;
 import org.sghweb.controllers.exceptions.RollbackFailureException;
@@ -31,7 +34,7 @@ public class ActividadJpaController implements Serializable {
 
     public ActividadJpaController(UserTransaction utx, EntityManagerFactory emf) {
         this.utx = utx;
-        this.emf = emf;
+        this.emf = emf; 
     }
     @Resource
     private UserTransaction utx = null;
@@ -46,13 +49,31 @@ public class ActividadJpaController implements Serializable {
     }
 
     public void create(Actividad actividad) throws PreexistingEntityException, RollbackFailureException, Exception {
-        EntityManager em = null;
+        if (actividad.getTurnoList() == null) {
+            actividad.setTurnoList(new ArrayList<Turno>());
+        }
+        EntityManager em = null;    
         Context initCtx = new InitialContext(); 
         utx = (UserTransaction) initCtx.lookup("java:comp/UserTransaction");
         try {
             utx.begin();
             em = getEntityManager();
+            List<Turno> attachedTurnoList = new ArrayList<Turno>();
+            for (Turno turnoListTurnoToAttach : actividad.getTurnoList()) {
+                turnoListTurnoToAttach = em.getReference(turnoListTurnoToAttach.getClass(), turnoListTurnoToAttach.getTurnoPK());
+                attachedTurnoList.add(turnoListTurnoToAttach);
+            }
+            actividad.setTurnoList(attachedTurnoList);
             em.persist(actividad);
+            for (Turno turnoListTurno : actividad.getTurnoList()) {
+                Actividad oldActividadOfTurnoListTurno = turnoListTurno.getActividad();
+                turnoListTurno.setActividad(actividad);
+                turnoListTurno = em.merge(turnoListTurno);
+                if (oldActividadOfTurnoListTurno != null) {
+                    oldActividadOfTurnoListTurno.getTurnoList().remove(turnoListTurno);
+                    oldActividadOfTurnoListTurno = em.merge(oldActividadOfTurnoListTurno);
+                }
+            }
             utx.commit();
         } catch (Exception ex) {
             try {
@@ -73,12 +94,47 @@ public class ActividadJpaController implements Serializable {
         }
     }
 
-    public void edit(Actividad actividad) throws NonexistentEntityException, RollbackFailureException, Exception {
+    public void edit(Actividad actividad) throws IllegalOrphanException, NonexistentEntityException, RollbackFailureException, Exception {
         EntityManager em = null;
+        Context initCtx = new InitialContext(); 
+        utx = (UserTransaction) initCtx.lookup("java:comp/UserTransaction");
         try {
             utx.begin();
             em = getEntityManager();
+            Actividad persistentActividad = em.find(Actividad.class, actividad.getCodigo());
+            List<Turno> turnoListOld = persistentActividad.getTurnoList();
+            List<Turno> turnoListNew = actividad.getTurnoList();
+            List<String> illegalOrphanMessages = null;
+            for (Turno turnoListOldTurno : turnoListOld) {
+                if (!turnoListNew.contains(turnoListOldTurno)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Turno " + turnoListOldTurno + " since its actividad field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<Turno> attachedTurnoListNew = new ArrayList<Turno>();
+            for (Turno turnoListNewTurnoToAttach : turnoListNew) {
+                turnoListNewTurnoToAttach = em.getReference(turnoListNewTurnoToAttach.getClass(), turnoListNewTurnoToAttach.getTurnoPK());
+                attachedTurnoListNew.add(turnoListNewTurnoToAttach);
+            }
+            turnoListNew = attachedTurnoListNew;
+            actividad.setTurnoList(turnoListNew);
             actividad = em.merge(actividad);
+            for (Turno turnoListNewTurno : turnoListNew) {
+                if (!turnoListOld.contains(turnoListNewTurno)) {
+                    Actividad oldActividadOfTurnoListNewTurno = turnoListNewTurno.getActividad();
+                    turnoListNewTurno.setActividad(actividad);
+                    turnoListNewTurno = em.merge(turnoListNewTurno);
+                    if (oldActividadOfTurnoListNewTurno != null && !oldActividadOfTurnoListNewTurno.equals(actividad)) {
+                        oldActividadOfTurnoListNewTurno.getTurnoList().remove(turnoListNewTurno);
+                        oldActividadOfTurnoListNewTurno = em.merge(oldActividadOfTurnoListNewTurno);
+                    }
+                }
+            }
             utx.commit();
         } catch (Exception ex) {
             try {
@@ -101,7 +157,7 @@ public class ActividadJpaController implements Serializable {
         }
     }
 
-    public void destroy(String id) throws NonexistentEntityException, RollbackFailureException, Exception {
+    public void destroy(String id) throws IllegalOrphanException, NonexistentEntityException, RollbackFailureException, Exception {
         EntityManager em = null;
         try {
             utx.begin();
@@ -112,6 +168,17 @@ public class ActividadJpaController implements Serializable {
                 actividad.getCodigo();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The actividad with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<Turno> turnoListOrphanCheck = actividad.getTurnoList();
+            for (Turno turnoListOrphanCheckTurno : turnoListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Actividad (" + actividad + ") cannot be destroyed since the Turno " + turnoListOrphanCheckTurno + " in its turnoList field has a non-nullable actividad field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(actividad);
             utx.commit();
@@ -174,9 +241,9 @@ public class ActividadJpaController implements Serializable {
             em.close();
         }
     }
-    
+        
     public List<Actividad> findActividadByDescripcion(String descripcion) {
         EntityManager em = getEntityManager();
         return em.createNamedQuery("Actividad.findByDescripcion").setParameter("descripcion", descripcion).getResultList();
-     }    
+    }   
 }
